style: ruff 指摘の解消 (不要な int() 2件 + import 順)
CI が ruff を回すため、このままだと push で落ちる状態だった。
いずれも形式のみでロジックへの影響は無い。

+++ backend/app/api/tide.py
@@ -225,7 +225,7 @@ def _habitual_bed_minute(session) -> int | None:
         h = circular_mean_hour(beds)
         if h is None:
             return None
-        return int(round(h * 60)) % 1440
+        return round(h * 60) % 1440
     except Exception:
         return None
